[FIX]hr_employee_calendar_planning:incorrect hours_per_day
Always using the topmost hr.employee.calendar for hours per day may lead to unexpected and incorrect behaviour. For example if the date_start and date_end of the topmost calendar are in the future. Then the hours_per_day from the future calendar are taken not from the currently used one. The more specific calendars in regard of date_start and date_end are considered earlier than the less specific ones.